% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_continuous}
\alias{scale_continuous}
\alias{scale_color_mn_c}
\alias{scale_colour_mn_c}
\alias{scale_fill_mn_c}
\title{Continuous scales to use for ggplot2}
\usage{
scale_color_mn_c(
  diverge = FALSE,
  high_index = "minnesota_blue",
  low_index = "minnesota_green",
  color_range = 0.75,
  ...
)

scale_colour_mn_c(
  diverge = FALSE,
  high_index = "minnesota_blue",
  low_index = "minnesota_green",
  color_range = 0.75,
  ...
)

scale_fill_mn_c(
  diverge = FALSE,
  high_index = "minnesota_blue",
  low_index = "minnesota_green",
  color_range = 0.75,
  ...
)
}
\arguments{
\item{diverge}{logical value to indicate whether color values should diverge.}

\item{high_index}{character color name for high values (e.g., "minnesota_blue"). See `show_palette_mn` for all names.}

\item{low_index}{character color name for low values (e.g., "minnesota_green").}

\item{color_range}{numeric specifying the amount of lightening. Negative numbers cause darkening.}

\item{...}{Arguments to pass on to `ggplot2::scale_color_gradient` (non-diverging) or
`ggplot2::scale_color_gradient2` (diverging)}
}
\value{
A `ScaleContinous` object that can be added to a `ggplot` object
}
\description{
These functions provide the option to use State of Minnesota continuous palette colors
with ggplot2.
}
\details{
Available State of Minnesota color names are available in the `show_palette_mn` function.
}
\examples{

if (require("ggplot2")) {
  ggplot(mtcars, aes(x = mpg, y = hp, color = wt)) +
    geom_point(size = 4) +
    scale_color_mn_c()
}
}
