% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_discrete}
\alias{scale_discrete}
\alias{scale_color_mn_d}
\alias{scale_colour_mn_d}
\alias{scale_fill_mn_d}
\title{Discrete scales to use for ggplot2}
\usage{
scale_color_mn_d(...)

scale_colour_mn_d(...)

scale_fill_mn_d(...)
}
\arguments{
\item{...}{Arguments to pass on to `ggplot2::discrete_scale()`}
}
\value{
A `ScaleDiscrete` object that can be added to a `ggplot` object
}
\description{
These functions provide the option to use State of Minnesota discrete palette colors
with ggplot2. There are ten available discrete colors available.
}
\details{
Available State of Minnesota color names are available in the `show_palette_mn` function.
}
\examples{

if (require("ggplot2")) {
  ggplot(mtcars, aes(x = mpg, y = hp, color = as.factor(carb))) +
    geom_point(size = 4) +
    scale_color_mn_d()
}
}
