% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_palette.R
\name{show_palette}
\alias{show_palette}
\alias{show_palette_mn}
\title{Show Palette}
\usage{
show_palette_mn()
}
\value{
A `ggplot` object
}
\description{
Display available State of MN named palette colors
}
